+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 3.1.0
 
+- Re-add `--fix` flag since some rules in standard (e.g. `lines-between-class-members`) are fixable, but not enforced by `prettier`. Such is the price for trying to bridge the two tools.
 - Upgrade to eslint 6.3.0 and latest standard, the one that fixes `no-unused-vars` false positives
 - Upgrade all other dependencies